+++ api/src/errors/customErrors.ts
@@ -25,7 +25,7 @@ export class UserNotFoundError extends APICustomError {
 
 export class PostNotFoundError extends APICustomError {
   constructor(identifier: string) {
-    super(`Post with the ID, ${identifier} not found.`, 404);
+    super(`Post with the ID, "${identifier}" not found.`, 404);
   }
 }