+++ api/src/validators/params.validator.ts
@@ -1,29 +1,80 @@
 import { RequestHandler } from "express";
 import asyncHandler from "express-async-handler";
 import { checkSchema, Schema, validationResult } from "express-validator";
-import { getUser } from "../services/user";
 import { BadRequestError, UserNotFoundError } from "../errors/customErrors";
 
-const isIdValid = async (id: string) => {
-  /* const userExists = !!(await getUser(id));
-  if (!userExists)
-    return Promise.reject(`User with the id, ${id}, does not exist`); */
-  try {
-    await getUser(id);
-  } catch (err) {
-    if (err instanceof UserNotFoundError) {
-      console.log("UserNotFoundError");
-      return Promise.reject(`User with the ID, ${id}, does not exist.`);
-    }
-    throw err;
-  }
-};
+interface ValidDataPostSchema {
+  username: string;
+  postPublicId: number;
+  postSlugTitle: string;
+}
 
-const userSchema = {
+const isPostSlugTitleValid = async (postSlugTitle: string) => {};
+
+const userSchema: Schema = {
   userId: {
     trim: true,
-    custom: {
-      options: isIdValid,
+    notEmpty: {
+      bail: true,
+    },
+    escape: true,
+  },
+};
+
+// TODO
+// custom validation for username and postSlugTitle
+const postSchema: Schema = {
+  username: {
+    trim: true,
+    notEmpty: {
+      errorMessage: 'Parameter "username" cannot be left empty',
+      bail: true,
+    },
+    escape: true,
+  },
+  postPublicId: {
+    trim: true,
+    notEmpty: {
+      errorMessage: 'Parameter "postPublicId" cannot be left empty',
+      bail: true,
+    },
+    isNumeric: {
+      errorMessage: 'Parameter "postPublicId" must be numeric',
+      options: {
+        no_symbols: true,
+      },
+    },
+    optional: {
+      options: {
+        nullable: true,
+      },
+    },
+    escape: true,
+    toInt: true,
+  },
+  postSlugTitle: {
+    trim: true,
+    notEmpty: {
+      errorMessage: 'Parameter "username" cannot be left empty',
+      bail: true,
+    },
+    optional: {
+      options: {
+        nullable: true,
+      },
+    },
+    escape: true,
+  },
+  commentId: {
+    trim: true,
+    notEmpty: {
+      errorMessage: 'Parameter "commentId" cannot be left empty',
+      bail: true,
+    },
+    optional: {
+      options: {
+        nullable: true,
+      },
     },
     escape: true,
   },
@@ -36,6 +87,9 @@ const validateParams = (schema: Schema): RequestHandler => {
       const errors = validationResult(req);
       if (!errors.isEmpty()) {
         // next({ status: "fail", code: 404, errors: errors.mapped() });
+        console.group("validateParams error thrown...");
+        console.log("errors.mapped():", errors.mapped());
+        console.groupEnd();
         throw new BadRequestError(
           "Validation failed for request parameters",
           404,
@@ -50,4 +104,9 @@ const validateParams = (schema: Schema): RequestHandler => {
   return paramsValidator;
 };
 
-export { validateParams as default, userSchema };
+export {
+  validateParams as default,
+  userSchema,
+  postSchema,
+  ValidDataPostSchema,
+};