+++ cms/src/components/navAnchor/NavAnchor.tsx
@@ -9,18 +9,13 @@ import { NavLink } from "react-router";
 interface Props<T = void> {
   pathname: string;
   textContent: string;
-  state?: T;
 }
 
 /* How to use generics for props in React functional component?
- * https://stackoverflow.com/a/66052574
+ * https://stackoverflow.com/questions/53958028/how-to-use-generics-in-props-in-react-in-a-functional-component/66052574#66052574
  */
-const NavAnchor = <T,>({ pathname, textContent, state }: Props<T>) => {
-  return (
-    <NavLink to={`${pathname}`} state={state}>
-      {textContent}
-    </NavLink>
-  );
+const NavAnchor = <T,>({ pathname, textContent }: Props<T>) => {
+  return <NavLink to={`${pathname}`}>{textContent}</NavLink>;
 };
 
 export default NavAnchor;